improved confirmation to delete
only for folders and allowing several nodes

+++ examples/megacmd/megacmdexecuter.cpp
@@ -96,7 +96,6 @@ MegaCmdExecuter::MegaCmdExecuter(MegaApi *api, MegaCMDLogger *loggerCMD)
     this->api = api;
     this->loggerCMD = loggerCMD;
     cwd = UNDEF;
-    nodeToConfirmDelete = NULL;
     fsAccessCMD = new MegaFileSystemAccess();
     mtxSyncMap.init(false);
     session = NULL;
@@ -105,7 +104,11 @@ MegaCmdExecuter::~MegaCmdExecuter()
 {
     delete fsAccessCMD;
     delete []session;
-    delete nodeToConfirmDelete;
+    for (std::vector< MegaNode * >::iterator it = nodesToConfirmDelete.begin(); it != nodesToConfirmDelete.end(); ++it)
+    {
+        delete *it;
+    }
+    nodesToConfirmDelete.clear();
 }
 
 // list available top-level nodes and contacts/incoming shares
@@ -2012,20 +2015,56 @@ int MegaCmdExecuter::actUponCreateFolder(SynchronousRequestListener *srl, int ti
     return 2;
 }
 
-
-
+/**
+ * @brief MegaCmdExecuter::confirmDelete
+ * @return number of elements to confirm/discard left
+ */
 void MegaCmdExecuter::confirmDelete()
 {
-    doDeleteNode(nodeToConfirmDelete,api);
-    nodeToConfirmDelete = NULL;
+    if (nodesToConfirmDelete.size())
+    {
+        MegaNode * nodeToConfirmDelete = nodesToConfirmDelete.front();
+        nodesToConfirmDelete.erase(nodesToConfirmDelete.begin());
+        doDeleteNode(nodeToConfirmDelete,api);
+    }
+
+
+    if (nodesToConfirmDelete.size())
+    {
+        string newprompt("Are you sure to delete ");
+        newprompt+=nodesToConfirmDelete.front()->getName();
+        newprompt+=" ? (Yes/No): ";
+        setprompt(AREYOUSURETODELETE,newprompt);
+    }
+    else
+    {
+        setprompt(COMMAND);
+    }
+
 }
 
+/**
+ * @brief MegaCmdExecuter::discardDelete
+ * @return number of elements to confirm/discard left
+ */
 void MegaCmdExecuter::discardDelete()
 {
-    nodeToConfirmDelete = NULL;
+    if (nodesToConfirmDelete.size()){
+        delete nodesToConfirmDelete.front();
+        nodesToConfirmDelete.erase(nodesToConfirmDelete.begin());
+    }
+    if (nodesToConfirmDelete.size())
+    {
+        string newprompt("Are you sure to delete ");
+        newprompt+=nodesToConfirmDelete.front()->getName();
+        newprompt+=" ? (Yes/No): ";
+        setprompt(AREYOUSURETODELETE,newprompt);
+    }
+    else
+    {
+        setprompt(COMMAND);
+    }
 }
-
-
 
 void MegaCmdExecuter::doDeleteNode(MegaNode *nodeToDelete,MegaApi* api)
 {
@@ -2061,10 +2100,32 @@ void MegaCmdExecuter::deleteNode(MegaNode *nodeToDelete, MegaApi* api, int recur
     }
     else
     {
-        if (interactiveThread() && !force)
+        if (interactiveThread() && !force && nodeToDelete->getType() == MegaNode::TYPE_FOLDER)
         {
-            setprompt(AREYOUSURETODELETE);
-            nodeToConfirmDelete = nodeToDelete;
+            bool alreadythere = false;
+            for (std::vector< MegaNode * >::iterator it = nodesToConfirmDelete.begin(); it != nodesToConfirmDelete.end(); ++it)
+            {
+                if (((MegaNode*)*it)->getHandle() == nodeToDelete->getHandle())
+                {
+                    alreadythere= true;
+                }
+            }
+            if (!alreadythere)
+            {
+                nodesToConfirmDelete.push_back(nodeToDelete);
+                if (getprompt()!=AREYOUSURETODELETE)
+                {
+                    string newprompt("Are you sure to delete ");
+                    newprompt+=nodeToDelete->getName();
+                    newprompt+=" ? (Yes/No): ";
+                    setprompt(AREYOUSURETODELETE,newprompt);
+                }
+            }
+            else
+            {
+                delete nodeToDelete;
+            }
+
             return;
         }
         else
@@ -2738,6 +2799,17 @@ void MegaCmdExecuter::executecommand(vector<string> words, map<string, int> *clf
         }
         if (words.size() > 1)
         {
+            if (interactiveThread() && nodesToConfirmDelete.size())
+            {
+                //clear all previous nodes to confirm delete (could have been not cleared in case of ctrl+c)
+                for (std::vector< MegaNode * >::iterator it = nodesToConfirmDelete.begin(); it != nodesToConfirmDelete.end(); ++it)
+                {
+                    delete *it;
+                }
+                nodesToConfirmDelete.clear();
+            }
+
+
             for (u_int i = 1; i < words.size(); i++)
             {
                 unescapeifRequired(words[i]);
